Return bfloat16 from bf16_approx_matmul and fix large-matrix correctness

+++ bitsandbytes/_ops.py
@@ -369,7 +369,7 @@ torch.library.define(
 def _(A: torch.Tensor, B: torch.Tensor, M: int, N: int, K: int) -> torch.Tensor:
     torch._check(A.dtype == torch.bfloat16, lambda: f"A must be bfloat16, got {A.dtype}")
     torch._check(B.dtype == torch.bfloat16, lambda: f"B must be bfloat16, got {B.dtype}")
-    return torch.empty((M, N), dtype=torch.float32, device=A.device)
+    return torch.empty((M, N), dtype=torch.bfloat16, device=A.device)
 
 torch.library.define(
     "bitsandbytes::bf16_approx_ewmul",
